+++ doc/html/search/all_2.js
@@ -1,0 +1,8 @@
+var searchData=
+[
+  ['addlevelbuttons',['addLevelButtons',['../classDanySpin97_1_1PhpBotFramework_1_1InlineKeyboard.html#a1a426eb2c9f438898fece4173ab06eb5',1,'DanySpin97::PhpBotFramework::InlineKeyboard']]],
+  ['addmessagecommand',['addMessageCommand',['../group__Bot.html#ga51215f7e30f84fa3148c4afe22de24a3',1,'DanySpin97::PhpBotFramework::Bot']]],
+  ['answerinlinequeryref',['answerInlineQueryRef',['../group__Optmized.html#gada4ab8135922ff6ffb1dd76b5726c004',1,'DanySpin97::PhpBotFramework::Bot']]],
+  ['api_20methods',['Api Methods',['../group__Api.html',1,'']]],
+  ['apirequest',['apiRequest',['../group__Api.html#gadf67b75ab357f9f4a0572b849b77267b',1,'DanySpin97::PhpBotFramework::CoreBot']]]
+];