+++ doc/html/search/all_3.js
@@ -1,0 +1,6 @@
+var searchData=
+[
+  ['bot',['Bot',['../classDanySpin97_1_1PhpBotFramework_1_1Bot.html',1,'DanySpin97::PhpBotFramework']]],
+  ['bot',['Bot',['../group__Bot.html',1,'']]],
+  ['botexception',['BotException',['../classDanySpin97_1_1PhpBotFramework_1_1BotException.html',1,'DanySpin97::PhpBotFramework']]]
+];