+++ doc/html/search/all_5.js
@@ -1,0 +1,7 @@
+var searchData=
+[
+  ['editinlinemessagetext',['editInlineMessageText',['../group__Api.html#ga2c42956774b0bfd50abf9e098b5c6069',1,'DanySpin97::PhpBotFramework::CoreBot']]],
+  ['editmessagetext',['editMessageText',['../group__Api.html#gad593230d6141a8fe77a868c29b124586',1,'DanySpin97::PhpBotFramework::CoreBot']]],
+  ['editmessagetextkeyboard',['editMessageTextKeyboard',['../group__Optmized.html#ga1a68379b200aff60ca97ff30007ace79',1,'DanySpin97::PhpBotFramework::Bot']]],
+  ['exec_5fcurl_5frequest',['exec_curl_request',['../group__Core.html#gaa96cfc1c13b556f5f81953a7de014842',1,'DanySpin97::PhpBotFramework::CoreBot']]]
+];