+++ doc/html/search/all_7.js
@@ -1,0 +1,22 @@
+var searchData=
+[
+  ['get',['get',['../classDanySpin97_1_1PhpBotFramework_1_1InlineKeyboard.html#a3eef7ca7625399a6338bbf1065c9c0f4',1,'DanySpin97::PhpBotFramework::InlineKeyboard']]],
+  ['getarray',['getArray',['../classDanySpin97_1_1PhpBotFramework_1_1InlineKeyboard.html#a24b5dd3cd099e802164b8b2690cf2b59',1,'DanySpin97::PhpBotFramework::InlineKeyboard']]],
+  ['getbackbutton',['getBackButton',['../classDanySpin97_1_1PhpBotFramework_1_1InlineKeyboard.html#ae3079d11eca9b2d2f85a9ef1cda75249',1,'DanySpin97::PhpBotFramework::InlineKeyboard']]],
+  ['getbackskipkeyboard',['getBackSkipKeyboard',['../classDanySpin97_1_1PhpBotFramework_1_1InlineKeyboard.html#a63800be28692c118eb8d25133d00db8a',1,'DanySpin97::PhpBotFramework::InlineKeyboard']]],
+  ['getchat',['getChat',['../group__Api.html#gafa82cdd34831bb7e15cefb6471763755',1,'DanySpin97::PhpBotFramework::CoreBot']]],
+  ['getchatid',['getChatID',['../group__Bot.html#ga2c33b57fdccb63b5bc3cb6d8f9b793e4',1,'DanySpin97::PhpBotFramework::Bot']]],
+  ['getchooselanguagekeyboard',['getChooseLanguageKeyboard',['../classDanySpin97_1_1PhpBotFramework_1_1InlineKeyboard.html#a44f577021e30662996b4cbaf4b334e32',1,'DanySpin97::PhpBotFramework::InlineKeyboard']]],
+  ['getdata',['getData',['../group__Bot.html#gacb883f40f30fdac64279d13f6a251be4',1,'DanySpin97::PhpBotFramework::Bot']]],
+  ['getlanguagedatabase',['getLanguageDatabase',['../group__Multilanguage.html#gaec7d8c26f938ce49e00680b43b109b43',1,'DanySpin97::PhpBotFramework::Bot']]],
+  ['getlanguageredis',['getLanguageRedis',['../group__Multilanguage.html#ga3db32461d6e58364afcf5b001c74a331',1,'DanySpin97::PhpBotFramework::Bot']]],
+  ['getlanguageredisascache',['getLanguageRedisAsCache',['../group__Multilanguage.html#ga5ade68539e290d934d9a02395474dd8d',1,'DanySpin97::PhpBotFramework::Bot']]],
+  ['getquery',['getQuery',['../group__Bot.html#gaa6b9913d4dc292753d1897255f14a093',1,'DanySpin97::PhpBotFramework::Bot']]],
+  ['getresults',['getResults',['../group__InlineQueryResults.html#gaf4c921b7b1b2561be3779055012e20f1',1,'DanySpin97::PhpBotFramework::InlineQueryResults']]],
+  ['getstatus',['getStatus',['../group__State.html#ga4ac793fc1611fe32ea7633d20fe30f39',1,'DanySpin97::PhpBotFramework::Bot']]],
+  ['gettext',['getText',['../group__Bot.html#ga285c76c1fb69800332147d5f7fc8aae9',1,'DanySpin97::PhpBotFramework::Bot']]],
+  ['getupdates',['getUpdates',['../group__Api.html#ga4c942ff7fbd47cc4b5536a0b7c16f4b1',1,'DanySpin97::PhpBotFramework::CoreBot']]],
+  ['getupdatesdatabase',['getUpdatesDatabase',['../group__Bot.html#gab715326bb3b458ebf8dbda635f2c8520',1,'DanySpin97::PhpBotFramework::Bot']]],
+  ['getupdateslocal',['getUpdatesLocal',['../group__Bot.html#gafbd201d6534fbba7aaae4ba42d9ace81',1,'DanySpin97::PhpBotFramework::Bot']]],
+  ['getupdatesredis',['getUpdatesRedis',['../group__Bot.html#ga98083d5b9fd47fc57571ceb9be4edb66',1,'DanySpin97::PhpBotFramework::Bot']]]
+];