+++ doc/html/search/all_8.js
@@ -1,0 +1,7 @@
+var searchData=
+[
+  ['inlinekeyboard',['InlineKeyboard',['../classDanySpin97_1_1PhpBotFramework_1_1InlineKeyboard.html',1,'DanySpin97::PhpBotFramework']]],
+  ['inlinekeyboard',['InlineKeyboard',['../group__InlineKeyboard.html',1,'']]],
+  ['inlinequeryresults',['InlineQueryResults',['../group__InlineQueryResults.html',1,'']]],
+  ['inlinequeryresults',['InlineQueryResults',['../classDanySpin97_1_1PhpBotFramework_1_1InlineQueryResults.html',1,'DanySpin97::PhpBotFramework']]]
+];