+++ doc/html/search/all_d.js
@@ -1,0 +1,14 @@
+var searchData=
+[
+  ['sendaudio',['sendAudio',['../group__Api.html#gaf4db6414a75197bccb73d66b18d1b396',1,'DanySpin97::PhpBotFramework::CoreBot']]],
+  ['sendchataction',['sendChatAction',['../group__Api.html#gad148dd91929db4bbfaeeb20f6df4c8bb',1,'DanySpin97::PhpBotFramework::CoreBot']]],
+  ['senddocument',['senddocument',['../group__Api.html#gaf0178e2d189cb613f5f58bb5982220d5',1,'DanySpin97::PhpBotFramework::CoreBot']]],
+  ['sendmessage',['sendMessage',['../group__Api.html#ga858b3c4333fc3bd11bc8d4cf7fe16e8b',1,'DanySpin97::PhpBotFramework::CoreBot']]],
+  ['sendmessagekeyboard',['sendMessageKeyboard',['../group__Optmized.html#ga8d714fa08899cc9819e1c93fcb732ae1',1,'DanySpin97::PhpBotFramework::Bot']]],
+  ['sendphoto',['sendPhoto',['../group__Api.html#ga8ec6e1bb18abb36994bdef01981d8ff3',1,'DanySpin97::PhpBotFramework::CoreBot']]],
+  ['sendsticker',['sendSticker',['../group__Api.html#gabd8394b0cdf205340b8e560240a02619',1,'DanySpin97::PhpBotFramework::CoreBot']]],
+  ['setchatid',['setChatID',['../group__Bot.html#gae5b20199df1cd6464a6ce90d9569af56',1,'DanySpin97::PhpBotFramework::Bot']]],
+  ['setlanguageredisascache',['setLanguageRedisAsCache',['../group__Multilanguage.html#ga64ed17b2a28fd265bd7da9c48bdf8d9e',1,'DanySpin97::PhpBotFramework::Bot']]],
+  ['setstatus',['setStatus',['../group__State.html#ga62343b2f096c8fc10d262b95765b74a8',1,'DanySpin97::PhpBotFramework::Bot']]],
+  ['state',['State',['../group__State.html',1,'']]]
+];