+++ doc/html/search/classes_1.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['corebot',['CoreBot',['../classDanySpin97_1_1PhpBotFramework_1_1CoreBot.html',1,'DanySpin97::PhpBotFramework']]]
+];