+++ doc/html/search/functions_2.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['clearkeyboard',['clearKeyboard',['../classDanySpin97_1_1PhpBotFramework_1_1InlineKeyboard.html#af00c350e57fbabf5ad7d111c73dd2b11',1,'DanySpin97::PhpBotFramework::InlineKeyboard']]]
+];