+++ doc/html/search/functions_4.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['forwardmessage',['forwardMessage',['../group__Api.html#ga2fcdc5f13f11b399d2c073da04815394',1,'DanySpin97::PhpBotFramework::CoreBot']]]
+];