+++ doc/html/search/functions_6.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['newarticle',['newArticle',['../group__InlineQueryResults.html#gad26b9ec605821b00c1efde2a030581a0',1,'DanySpin97::PhpBotFramework::InlineQueryResults']]]
+];