+++ doc/html/search/groups_3.js
@@ -1,0 +1,5 @@
+var searchData=
+[
+  ['inlinekeyboard',['InlineKeyboard',['../group__InlineKeyboard.html',1,'']]],
+  ['inlinequeryresults',['InlineQueryResults',['../group__InlineQueryResults.html',1,'']]]
+];