+++ doc/html/search/groups_5.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['optmized_20api_20methods',['Optmized Api Methods',['../group__Optmized.html',1,'']]]
+];