+++ doc/html/search/searchdata.js
@@ -1,0 +1,27 @@
+var indexSectionsWithContent =
+{
+  0: "$_abcefgimnops",
+  1: "bci",
+  2: "_acefgnps",
+  3: "$",
+  4: "abcimos"
+};
+
+var indexSectionNames =
+{
+  0: "all",
+  1: "classes",
+  2: "functions",
+  3: "variables",
+  4: "groups"
+};
+
+var indexSectionLabels =
+{
+  0: "All",
+  1: "Classes",
+  2: "Functions",
+  3: "Variables",
+  4: "Modules"
+};
+