+++ templates/data_example.php
@@ -1,8 +1,0 @@
-<?php
-
-$user = 'user';
-$password = 'pass';
-$token = 'token';
-$dbname = 'dbname';
-$driver = 'driver';
-$redis_address = '127.0.0.1';