+++ tests/bottest.php
@@ -7,9 +7,7 @@ class BotTest extends PHPUnit_Framework_TestCase
     public $bot;
 
     public function setUpBot() {
-        $this->bot = new \WiseDragonStd\HadesWrapper\Bot("token");
-        $this->bot->connectToRedis();
-        $this->bot->connectToDatabase("mysql", "hello_world_test", "root", "mysql");
+        $this->bot = new DanySpin97\PhpBotFramework\Bot("token");
 
     }
 }