+++ tests/keyboardtest.php
@@ -5,8 +5,8 @@ require './vendor/autoload.php';
 class KeyboardTest extends PHPUnit_Framework_TestCase
 {
     public function testInlineKeyboard() {
-        $keyboard = new \WiseDragonStd\HadesWrapper\InlineKeyboard();
+        $keyboard = new DanySpin97\PhpBotFramework\InlineKeyboard();
         $keyboard->addLevelButtons(['text' => 'First', 'callback_data' => '1'], ['text' => 'Second', 'callback_data' => '2']);
-        $keyboard->getKeyboard();
+        $keyboard->get();
     }
 }